+++ src/components/CreateOrder/stages/FinishMessage/FinishMessage.jsx
@@ -4,7 +4,7 @@ const FinishMessage = () => {
   const stage = useSelector((state) => state.form.currentStage);
   console.log(stage);
   const isHide = stage != 6
-  
+   
    return (
       isHide ? null : <div>Форма отправлена!</div>
    )
